requirements.txt
fastmcp
llama-index
llama-index-core
llama-index-embeddings-huggingface
httpx
pandas
streamlit
python-dotenv
huggingface_hub
mnest-asyncio
uvicorn